requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
arrow==0.7.0
google-api-python-client==1.4.2
httplib2==0.9.2
itsdangerous==0.24
oauth2client==1.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-dateutil==2.4.2
rsa==3.2.3
simplejson==3.8.1
six==1.10.0
uritemplate==0.6
urllib3==1.12
pymongo==3.0.3
gunicorn==19.4.5
